+++ crates/javy/CHANGELOG.md
@@ -8,7 +8,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [3.1.0] - 2024-11-26
+## [3.1.0] - 2024-11-27
 
 ### Added